Dockerfile: update criu stage to Debian 12 (Bookworm)
Follow-up to e72c4818c400efe0b0b942bfcbc244287cfdfbd3, which updated the
Dockerfile to use Debian 12 "bookworm", but forgot to update the package
repository to use for the CRIU packages. Note that the criu stage is currently
not built by default (see d3d2823edfb0247fe3ee320414dc3c836e53a3b9), so to
verify the stage, it needs to be built manually;

    docker build --target=criu .

This patch adds an extra `criu --version` to the build, so that it's verified
to be "functional".

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>