client: add support for overriding minimum API version
This adds options to override the minimum API version considered
when performing API-version negotiation. These options allow the
client to either downgrade to versions lower than supported (lower
than MinAPIVersion), which may result in degraded functionality,
or to raise the minimum version to exclude older API versions.

The `WithMinAPIVersion` can be used to set the version programmatically.
The `WithMinAPIVersionFromEnv` allows a user to set the minimum API
version at runtime through the `DOCKER_MIN_API_VERSION` env-var.

The `WithMinAPIVersionFromEnv` is more permissive to account for
slightly malformed valus in the environment variables; it accepts
whitespace, and API versions with a "v" prefix. It also ignores
empty values in the environment-variable (which are considered a
no-op).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>