pkg/ioutils: Fix possible double close in CancelReadCloser
Picking the `select` branch is not deterministic - after `ctx.Done` is
closed and `doneCtx.Done` gets closed before the `select` block is
reached again there's no guarantee that the `doneCtx` case will be
picked and the `closeWithError` may be invoked again.

Remove the loop and exit the goroutine as soon as any of these contexts
is cancelled. `doneCtx` will only be cancelled by the `closeWithError`
so there's no need to call it again.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>