c8d: ImageService.GetImage: fix filtering by platform
Before this patch, it would find all images (and platform) are found,
and return the first image (or the old default: "linux/amd64");

With this patch, an error is returned if a platform is set, and no
results were found:

    No such image: alpine:latest for platform: windows/s390x

Looking at consumers of this function, there may be something to fix though;
https://github.com/moby/moby/blob/178651733852a26bc11e6fa272b48ff37ecc7447/daemon/create.go#L85-L86

    if opts.params.Platform == nil && opts.params.Config.Image != "" {
        img, err := daemon.imageService.GetImage(ctx, opts.params.Config.Image, imagetypes.GetImageOpts{Platform: opts.params.Platform})

The code above looks to check for _NO_ platform to be passed in parameters, but
then to pass that platform as `imagetypes.GetImageOpts.Platform`.

That code was added in 300c11c7c9143ca077890050deb83917e67fe250 but that was a
refactor, combining multiple branches, so before that we passed platform
unconditionally (but only used the image in some cases, so that was an
optimization).

Should we;

- always lookup the image?
- should `GetImage` use OnlyPlatformWithFallback as default (if no platform
  is specified)? To select any image, but _prefer_ local platform?

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>