pkg/platforms: internalize in daemon/containerd
This matcher was only used internally in the containerd implementation of
the image store. Un-export it, and make it a local utility in that package
to prevent external use.

This package was introduced in 1616a09b613b07e56926ae23b1ae710f9a2c64fe
(v24.0), and there are no known external consumers of this package, so there
should be no need to deprecate / alias the old location.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit 94b476536342b89a78250fdf349539092db5dd27)
Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>