volume/local: Make host resolution backwards compatible
Commit 8ae94cafa50a90192a0c0189f54443abac7a9b44 added a DNS resolution
of the `device` part of the volume option.

The previous way to resolve the passed hostname was to use `addr`
option, which was handled by the same code path as the `nfs` mount type.

The issue is that `addr` is also an SMB module option handled by kernel
and passing a hostname as `addr` produces an invalid argument error.

To fix that, restore the old behavior to handle `addr` the same way as
before, and only perform the new DNS resolution of `device` if there is
no `addr` passed.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>
(cherry picked from commit 0d51cf9db82f196af28b7b69139bb01154c4f72a)
Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>