libnet: remove retry logic from endpointCnt.atomicIncDecEpCnt
This retry logic was originally added by moby/libnetwork@7393533 (PR
moby/libnetwork#626, merged on Oct 10, 2015), at a time when this method
was on `*network`. Then, a new `endpointCnt` struct was added by
moby/libnetwork@cd49ffd (PR moby/libnetwork#633, merged on Oct 12,
2015).

The second PR has another commit bundled (moby/libnetwork@c19a779), and
that commit's message gives a good clue about why the first PR was even
needed:

> Currently there are 3 distinct operations performed by
> datastore
>    - Pushing the data to the store
>    - Updating the Index of the local object
>    - Updating the cache (in case of localscope)
>
> Without a lock racing datastore api calls can interleave
> in various surprising ways. Best thing is to keep these
> 3 above operation inseparable. Use a datastore lock to
> achieve this.

Another thing to keep in mind, Classic Swarm was under active
development at that time  and the datastore API was already
supporting reading from / writing to eventually consistent
datastores.

So it looks like the retry logic was needed because of:

1. Distributed data stores with no strong consistency model ;
2. Requests to the local datastore not being serialized, leading
   to consistency issues ;

Both issues are a thing of the past.

But for issue 2. it's important to note that there is possibly
another contributing factor: datastore's `GetObject` method
was (and still is) hydrating the `KVObject` parameter passed as
argument instead of returning the object in the cache (if any).
As a consequence, two instances of the same object could live in
memory at the same. This would defeat the purpose of mutex,
and one instance could lag behind the other (in terms of dbIndex,
the version ID used for optimistic locking).

While the current design of the datastore doesn't disallow this
issue, a static analysis of the code shows that it doesn't happen.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>