libnet/d/bridge: remove retry logic from driver.storeDelete
This retry logic was added by moby/libnetwork@845ef40. Neither
commit messge nor PR description indicates why this is needed.
Like other commit in this series, this was done to support
Classic Swarm and its eventually consistent datastores.

Now that we don't support these datastores anymore, the only
case where this logic could be somewhat useful is if two
separate instances of the same objects and with a different
dbIndex live in memory at the same time. This could happen
if the driver was calling `d.store.GetObject()`. Fortunately,
the driver doesn't do that in any other code path, so that's
a non-issue.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>