libnet/d/[ip|mac]vlan: remove retry logic from driver.storeDelete
The ipvlan / macvlan drivers were added in 1d6f2c5, submitted Feb 23,
2016. The original commit already had the retry logic. Neither
the commit message nor the PR mention why this is needed.

But since this driver was added after the retry logic was added to
the bridge driver, there's a high probability this was carried over
from there. Whether this was useful at all, I don't know. But...

As we got rid of the external datastores, these are mostly unneeded.
However, there's one possible case: two separate instances of the
same object with a different dbIndex living in memory. The only way
to end up with that is if the driver calls `d.store.GetObject()`.
Fortunately, the driver doesn't do that in any other path so that's
a non-issue.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>