libnet: remove the `EndpointInfo` interface
`EndpointInfo` was introduced in moby/libnetwork@d8ba1e231. The commit's
message states:

> This sort of deisgn hides the libnetwork specific type details from
> drivers.

9 years later it turns out the `Endpoint` struct is the only implementor
of that interface, and `libnetwork` and `daemon` packages are the only
consumers. This is adding an unneeded level of indirection.

`(*Endpoint).Info()` is the only function that returns an `EndpointInfo`.
It's basically 12 lines of cursed indirection packed together. But, as
noted by thaJeztah in 210abfa, we need to make sure the `*Endpoint`
receiver is always fully hydrated to get rid of it. So here we go.

As the diff shows, most of the time `(*Endpoint).Info()` is called right
after a call to `(*Network).Endpoints()` (sometimes through an
indirection). That method in turns calls `getEndpointsFromStore`, which
is calling `(*Store).List()`, which is either retrieving all objects
of a given type from its internal cache (if it was populated), or
retrieve them all from boltdb. In that case, `*Endpoint`s are fully
hydrated.

- `disconnectFromNetwork`: calls `WalkEndpoints`, which calls
  `(*Network).Endpoints()`, right before ;
- `buildContainerAttachments`: calls `(*Network).Endpoints()` right before ;
- `buildEndpointResource`: `ep` and `info` are actually both the same thing ;
- `clearAttachableNetworks`: calls `(*Network).Endpoints()` right before ;
- `findLBEndpointSandbox`: calls `(*Network).Endpoints()` right before ;
- `addLBBackend`: calls `(*Network).Endpoints()` right before ;
- `deleteLoadBalancerSandbox`: calls `(*Network).EndpointByName()`, which
  calls `WalkEndpoint` (see above).

We're left with the following, less obvious cases:

- `updateJoinInfo`: called from `(*Daemon).connectToNetwork()`, where
  `*Endpoint`s are created ;
- `buildEndpointInfo`: called from `(*Daemon).updateEndpointNetworkSettings()`,
  again called from `(*Daemon).connectToNetwork()` ;

Meaning, we can now safely assume `*Endpoint`s are always fully hydrated
before calling its `Info()` method. We can finally ditch it! :tada:

Signed-off-by: Albin Kerouanton <albinker@gmail.com>