libn/d/overlay/ovmanager: drop subnet struct
It is only a single uint32 wrapped in a struct; a pointer to the struct
is twice the size of the struct itself. Inline the list of VNIs
allocated to the network into the network struct.

Drop the slice when releasing the network's VNIs instead of zeroing out
the slice members.

Signed-off-by: Cory Snider <csnider@mirantis.com>